src/utils/requirements.txt
joblib==1.5.1
langchain==0.3.27
langchain_chroma==0.2.5
langchain_community==0.3.27
langchain_groq==0.3.6
langchain_huggingface==0.3.1
python-dotenv==1.1.1
